Break out projectile file
